--- src/delete/actionTypes.js
@@ -1,2 +1,0 @@
-export const LOGIN = 'LOGIN';
-export const LOGOUT = 'LOGOUT';
