--- src/delete/authActions.js
@@ -1,10 +1,0 @@
-import { LOGIN, LOGOUT } from "./actionTypes";
-
-
-export const login = () => ({
-    type:LOGIN
-});
-
-export const logout = () => ({
-    type:LOGOUT
-})
